cr: removed some warnings

+++ db4oj/core/src/com/db4o/DTrace.java
@@ -299,7 +299,7 @@ public class DTrace {
     
     public static void addRangeWithLength(long start, long length){
         if(enabled){
-            addRangeWithEnd(start, start + length - (long)1);
+            addRangeWithEnd(start, start + length - 1);
         }
     }
